Made edges less straight, something something trans agenda
Specifically implemented quadratic Bezier curves where the control point
is the point that was closest to making the edge not exist (i.e. closest
to the lens that must be empty for an edge to form).
Also changed how liquid placement works, it now produces nice edges of
water along the sides of caves.