Limit max y-level to near surface to improve speed
Now the cellular automata and digging only fill a much smaller array,
that is bounded by the chunk's surface level instead of the world build
height. I was worried this might introduce chunk discontinuities, but
because it already simulates a bit above the build height, it does not
actually cause them. This is because the rounds constant adds enough
padding that even near the surface there still aren't any
discontinuity-producing mismatches in what the cellular autonata's input
should be.