Line digging optimization
Now when digging a line (either as part of a regular cave or as part of
the individual lines that make up the inside of a triangle in a ravine),
it shrinks the loop's bounds so it doesn't iterate if the line is suitably
outside of the generating chunk. This however does not apply for the
initial edge used in digging triangles, since it is possible for an edge
of a triangle to be outside of the chunk while a part of the triangle as
a whole is.